+++ field/field-locator-contact.js
@@ -1,6 +1,6 @@
 
 const FieldLocator = require('./field-locator').FieldLocator;
-// const FieldTextBoolean = require('./field-text-boolean').FieldTextBoolean;
+const FieldTextBoolean = require('./field-text-boolean').FieldTextBoolean;
 const FieldLocatorText = require('./field-locator').FieldLocatorText;
 const FieldLocatorGuid = require('./field-locator').FieldLocatorGuid;
 const _ = require('lodash');
@@ -10,7 +10,6 @@ class FieldLocatorContact extends FieldLocator {
   constructor(options = {}) {
     super(options);
     this._fields = _.merge(this._fields, {
-      // _sync : new FieldTextBoolean(),
       id: new FieldLocatorGuid({ emptyAllow: false}),
       fullName: new FieldLocatorText({ emptyAllow: false}),
       name: new FieldLocatorGuid({emptyAllow: false}),
@@ -19,6 +18,10 @@ class FieldLocatorContact extends FieldLocator {
       search: new FieldLocatorText({ emptyAllow: false}),
       typeId: new FieldLocatorText({ emptyAllow: false}),
       type: new FieldLocatorText({ emptyAllow: false}),
+      // do a lookup on the email adress
+      email: new FieldLocatorText({emptyAllow: false}),
+      // if true and there are multiple parents the first one is used
+      _allowMulti: new FieldTextBoolean(({emptyAllow: false}))
     });
   }
 }