+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # adrez-mapper Revisions
 
+2020-06-16
+- add: locator-contact: email
+- add: locator-contact: _allowMulti
+
 2020-06-02
 - add: auto array for record elements
 - chg: country lookup param changes
+- add: test memo
 
 2020-06-29
 - add: name parser can be configured by a config. Keys: suffix, prefix, titles